21 keypad conundrum
Unfortunately^(n+1), a keypad cannot be used directly, so use a robot
arm to use a robot arm to use a robot arm to push a keypad. Moving a
robot arm with cartesian buttons obviously has two shortest paths
because bouncing back and forth would be the worst. Either take the x
movement and then the y movement or vice versa, followed by the A key.
The keypads have an unreachable corner though so don't choose paths that
would cross it. Moving from a key to A never crosses it. Try these
movements recursively to find the best key pressing cost until we're at
the human who stops the madness.

But, TODO: The robot arm controls always restart at A for each new
movement on the deepest pad (because they also end at A), so this could
be plenty simplified.