22b just the ones digit
Find the changes in price, and for each sequence of four changes, add
the possible price matching the sequence to the seq price if that's the
first of such sequence.

Single digit diffs can be in range from 0-9=-9 to 9-0=9 so the [diff; 4]
that in a way makes up a base-19 number is easily a space small enough
to have as a hashmap key.