23 lan party
Given a list of graph edges, find complete triplet subgraphs: Go through
each node x in the network and look at edges to form prev-x-next such
that also prev-next is a link. Also at least one of them has to start
with the letter t. Sort the triplets by name and put them in a set to
track the count of unique triplets.