23b go there yourself
Find the largest complete subgraph of the whole graph to identify the
LAN party.

Looking at the sample input and the quite consistent number of neighbors
in the real input, the LAN party probably has as many computers as those
nodes have neighbors, so cheat a bit with a solution that only works for
this easily-formed input: for each node N, for each neighbor M, observe
a subgraph of N and all its friends except M. Add one to the count of
that subgraph. Then after going through all nodes, the subgraph with the
largest count is the LAN party.

For a more general solution done this way we'd have to look at subgraphs
of not just one neighbor removed but all possible combinations; this
would be rather excessive and a better solution could be in order.

A picture of this case for demonstration, with example max size 4:

     ...    ...
       \    /
        +--+
        |\/|
        |/\|
        +--+
       /    \
     ...    ...