17: build high bits for last digits first
The high bits affect some first digits but not vice versa, so it makes
more sense to recursively compute the higher reg val options that
produce the end of the program (last digits) first. Then, on each
recursion back, go through those options and try the next (previous?)
three-bit variations.

This cuts runtime from ~500 ms to ~10 ms.

This recursion could also be just a loop; each deeper call happens only
once, so it's not a DFS problem anymore like it used to be.