24 crossed wires
Given a logic circuit as a system of gates and wires, simulate it from
the input state and see what number the z wire bits represent at the
end. The simulation is easy because there are no loops; the start state
is some signals on some input wires, and we just go through each gate
and ensure its output has been computed from the inputs, recursively.

(This could break if some intermediate wires were specified in the
input, giving the false impression that they're the final state.
Fortunately that's not the case. Another way could be to take all inputs
and recursively propagate their signals through all gates in a BFS-ish
fashion.)