24b inspect the monitoring device more closely
The circuit is a series of 1-bit adders but four pairs of gates have
their wires swapped. Trying each possible pair combination would take
forever, so let's instead look at the circuit and observe how things can
go wrong. Use the following heuristics to find bad wires and collect
them into the final list; the pair information is not necessary.

- a Z has to be an output of a XOR
- an AND must go to OR, except for the first X and Y
- an OR has to be an output of an AND
- an XOR inputting X or Y must output to a XOR that outputs to a Z
- an XOR outputting Z must input a XOR that inputs X or Y

This is sufficient at least for my input. If a wire is found that
contradicts with the above, that wire must be misplaced.