25 code chronicle
We have ascii art of locks and keys that probably could be verified on
its own, but parse locks and keys into series of pin sizes and key shape
heights and then verify each lock and key together for fit. This is not
exact fit though so the result is optimistic.