24: general circuit verification overengineered
Turn the hardcoded loops into generics-based chains of checks that
validate that a wire in some pattern must exist in the input. Typing and
reading the raw turbofish that many times is pretty jarring, so build
tooling to chain values of the check types, and abuse operator
overloading to make it look even nicer. The following returns an
iterator over all mismatched wires where a carry bit chain from a X
(except the first X) to a Z (whose number follows that of the X by 1) is
expected:

    Circuit::<X::<1, 43, And::<Or::<Xor::<Z::<1>>>>>>::verify_t(device)

And the same in a more readable form:

    x::<1, 43>() | and | or | xor | z::<1>() | verify(device)

Checks can even be chained:

    let xz    = x::<1, 44>() | xor |      xor | z::<0>();
    let carry = x::<1, 43>() | and | or | xor | z::<1>();
    let iter = xz + carry | verify(device);