3b overcome the static friction
Now the max joltage is composed of 12 batteries, not just two. That's
too many to bruteforce through. For each digit with N other digits of 12
remaining, find the max of 1) that digit and the same max of the next
with N-1, or 2) the next with N remaining, skipping that one. Memoize
the results to save on repetition.

This is rather slow, might be better to start from the other end or do
some heuristics on finding the greatest digit with still enough digits
on the right.