8b enough extension cables
Now the question is how many connections are needed until there is only
one circuit. Any connections after that are just extra. The first part's
logic gives results for n=1000, we don't know n, so reuse the logic and
binary search the correct n and take the password-like X coordinates of
the last two junction boxes connected.

This is an awful hack; a proper iterative solution shall be devised at
once.