9b the elves just remembered
Now not all tiles can contribute to the grand-area rectangle but only
those that are within the region specified by the given input that is
like fencepost coordinates. The area is not necessarily even convex. Go
over each pair again and filter out rectangles that would be cut by any
line in the input.

Somehow this works; the input happens to be sufficiently convex. With
e.g. a T-shaped region the "best" might still be found from the bottom
left or right but it couldn't be the best because the area would be
outside the valid region.