10 factory
The required set of indicator lights is small enough so represent them
as a bitmap. Each button toggles a set of lights so they too are bitmaps
of which lights to flip; flipping a state of all lights with a button is
now a simple bitwise xor. For the shortest path from blank lights to the
desired state, use the power of copypasta and slap a Dijkstra search on
it.

In hindsight, Dijkstra is probably excessive because all edge lengths
are equal; a plain BFS should do. Also, there are so few button
combinations that simply looping over 2^buttoncount and treating buttons
as bits in the loop index would be perfectly viable.

Most of the time on this went in debugging an input parsing bug, argh.