11 reactor
With an adjacency list for a network, find the number of different paths
between two specific nodes. Weirdly this feels very much like part two
of day 7. DFS from start to end gives 1 for the end node and 0 for a
node with no path, and the sum of paths from neighbors otherwise.
Memoizing takes care of the galactic scale of repetition.